third_party/WebKit/LayoutTests/mhtml/mhtml_in_iframe-expected.txt

CONSOLE ERROR: Attempted to load a multipart archive into an subframe: simple_page_unmht.mht
This page has an IFrame that tries to load some MHTML.
The load should fail.


--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
